refactor: move/delete utils and fix tests and imports
