+++ openai_assistant/utils/mailchimp.py
